WIP: more rearranging of packages.
* Rename the now public FunctionResult to AbstractFunctionResult
* Extrtact FunctionRsult from the abstract class as an interface.
