+++ geode-core/src/main/java/org/apache/geode/management/internal/cli/functions/CreateAsyncEventQueueFunction.java
@@ -39,7 +39,7 @@ import org.apache.geode.internal.ClassPathLoader;
 import org.apache.geode.internal.InternalDataSerializer;
 import org.apache.geode.internal.cache.InternalCache;
 import org.apache.geode.internal.logging.LogService;
-import org.apache.geode.management.FunctionResult.StatusState;
+import org.apache.geode.management.AbstractFunctionResult.StatusState;
 import org.apache.geode.management.cli.CliFunction;
 import org.apache.geode.management.cli.CliFunctionResult;
