+++ geode-core/src/main/java/org/apache/geode/management/internal/cli/functions/DestroyGatewayReceiverFunction.java
@@ -24,7 +24,7 @@ import org.apache.geode.cache.execute.FunctionContext;
 import org.apache.geode.cache.execute.ResultSender;
 import org.apache.geode.cache.wan.GatewayReceiver;
 import org.apache.geode.internal.logging.LogService;
-import org.apache.geode.management.FunctionResult.StatusState;
+import org.apache.geode.management.AbstractFunctionResult.StatusState;
 import org.apache.geode.management.cli.CliFunction;
 import org.apache.geode.management.cli.CliFunctionResult;
 import org.apache.geode.management.internal.cli.CliUtil;
